+++ webpack.config.dev.js
@@ -26,6 +26,10 @@ module.exports = {
             {
                 test:/\.s[ac]ss$/,
                 use:['style-loader','css-loader','sass-loader']
+            },
+            {
+                test:/\.(png|svg|jpg|jpeg|gif)$/,
+                type:'assets'
             }
         ]
     },
@@ -40,6 +44,7 @@ module.exports = {
     ],
     devServer: { // configuración para el server
         static: path.join(__dirname, 'dist'), // ruta de nuestro dist
+        historyApiFallback: true, // llevar histórico para la navegación
         compress: true, // si deseamos comprimir
         port: 3006, // el puerto que deseamos utilizar
         open: true, // para abrir nuestro navegar automáticamente