+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Change to ospinajuanp-single-page
 
-## 1.1.0 
+## 1.1.1 (January 29, 2023)
+* React Route 5 implementation
+* Copy of assets implementation
+* Folder structure implementation (Atom Design)
+    - components 
+    - containers 
+    - pages 
+    - routes
+    - style
+
+## 1.1.0 (January 28, 2023)
 
 * React implementation
 * Webpack implementation
@@ -11,7 +21,6 @@
 * Babel implementation
 * sass implementation
 * ospinajuanp-reset-css implementation
-* 
 
 ### 1.0.3 (January 28, 2023)